django_data/requirements.txt
django
djangorestframework
django-cors-headers
django-bootstrap4
django-allauth
django-rest-auth
django-filter
django-url-filter
django-crispy-forms
django-url-filter
djangorestframework-queryfields
pymysql
gunicorn
ptvsd
requests
urllib3
chardet
certifi
idna
numpy
six
mysqlclient
Pillow